pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "awesome-greek-software-engineering"
version = "0.0.0"
description = "Generator tooling for the Awesome Greek Software Engineering directory"
readme = "readme.md"
requires-python = ">=3.10"
dependencies = [
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "requests>=2.28",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
awesome_greek_software_engineering = ["py.typed"]
